requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
openai>=1.50.0
tavily-python>=0.5.0
python-dotenv>=1.0.0